tidy up organise files displays and comments
