+++ views/postDisp.ejs
@@ -4,7 +4,7 @@
         <a href=".."><h3>Back to previous page</h3></a>
     <h1>List of portfolios for user <%= path1 %> below</h1>
 
-    
+    <!-- LIST OF ALL PORTOLIOS WITH DELETE BUTTON -->
     <% portfolios.forEach(function(portName) { %>
       <div class="Row">
       <a href="/<%- path1 %>/<%- portName %>"<h2><%- portName %></h2></a>
@@ -18,8 +18,9 @@
      <% }) %>
 
      <br><br><br>
+     <!-- ADD PORTFOLIO BUTTON -->
      <form method="post" action="/addPortfolio">
-       <input type="text" name="portfolioToAdd" value="New_Portfolio_Name">
+       <input type="text" name="portfolioToAdd" value="New_Portfolio_Name" pattern="[0-9A-Za-z_]*">
        <!-- make it so quotes, spaces ect can't be input -->
        <input type="hidden" name="userName" value="<%= path1 %>">
        <input type="submit" value="Add Portfolio">
